[exporter] fixes a bug baking shadow texture don't work properly (#17)
The issue was caused by attempting to extract and bake
a null texture from `_bakedMaterialMainTextures`.
As a result, the main texture that should have been
assigned to the shadow texture was missing during
the baking process. This led to the generation of
a 2x2 white shadow texture, causing shadows to be
incorrectly rendered in MToon.

To fix this, `_bakedMaterialMainTextures` has been
modified to always contain the main texture.
Additionally, the information from
`gltf.material.TextureInfo` is now stored to prevent
redundant processing by `ExportTextureInfoMToon`.