[export] overhaul of texture processing (#18)
This commit includes two major changes.

1. Standardizing output to RGBA32 textures
Initially, to reduce texture size, RGBA32 was used when an alpha channel
was needed, while RGB24 was used when it was not. However, the added
complexity outweighed the benefits, so all texture output has been
unified to RGBA32. According to the glTF specification, unused channels
are simply ignored, so this change does not cause any issues.

2. Aligning texture storage with glTF
Previously, normal maps, metallic roughness textures, and occlusion
textures were stored as is, without addressing the differences in
storage methods between Unity and glTF. This has been corrected to match
glTF's format using the same approach as UnityGLTF.