[exporter] process PB branches as independent segments (#33)
* [exporter] process PB branches as independent segments

Previously, PB branches were not considered, resulting in a single continuous chain being exported as a Spring Bone. With this fix, branches are now handled separately, and when a split occurs, each branch is processed independently. However, since VRM does not define behavior for branched Spring Bones and relies on implementation specifics, this will be explicitly noted in the README.

* [exporter] fix an issue where empty chains were generated

Due to the nature of `RetrieveSpringBoneChainTransforms`, empty chains were being generated. This update removes them after the function call. Additionally, a bug where depth calculation incorrectly counted nodes even when no descendants existed has been fixed. Furthermore, when a spring bone chain branches, numbers are now appended to their names, and this behavior is documented in the `README.md`.