[exporter] Fixed multiple bugs in material output (#56)
This commit fixes a bug where `IndexOutOfRangeException` occurs when the number of submeshes is greater than the number of materials. In such cases, `sharedMaterial` result is applied to out-of-range submeshes, and information is output to the NDMF console.

I also confirmed that there is an issue where `TryGetValue` throws an exception when the material is `None`, so in such cases, information will be output to the NDMF console on a per-mesh basis and the processing will be skipped.